% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictCellType.R
\name{confidence_score}
\alias{confidence_score}
\title{Produce confidence score for XGBoost prediction}
\usage{
confidence_score(vec)
}
\arguments{
\item{vec}{A vector of predicted probability for each cell type}
}
\value{
A numeric confidence score from 0 to 1.
}
\description{
Produce confidence score for XGBoost prediction
}
