% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictCellType.R
\name{confidence_score_top95}
\alias{confidence_score_top95}
\title{Produce confidence score based on top 95 percent for XGBoost prediction}
\usage{
confidence_score_top95(vec)
}
\arguments{
\item{vec}{A vector of predicted probability for each cell type}
}
\value{
A numeric confidence score from 0 to 1.
}
\description{
Produce confidence score based on top 95 percent for XGBoost prediction
}
