% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictCellType.R
\name{filter_cell}
\alias{filter_cell}
\title{Filter final prediction to reduce noise}
\usage{
filter_cell(pred_result, knn_res, KNeighbor = 5)
}
\arguments{
\item{pred_result}{The prediction result from XGBoost}

\item{knn_res}{knn graph from smooth_matrix}

\item{KNeighbor}{Number of knn neighbors to use for smoothing (Default: 5)}
}
\value{
The final prediction result after dropping few cell types
}
\description{
Filter final prediction to reduce noise
}
