% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateInput.R
\name{GenerateInput}
\alias{GenerateInput}
\title{Generate pattern level data for cell type annotation}
\usage{
GenerateInput(query_fn, knowledge_fn)
}
\arguments{
\item{query_fn}{File path to query .cg}

\item{knowledge_fn}{File path to pattern file .cm}
}
\value{
A cell by pattern matrix.
}
\description{
Generate pattern level data for cell type annotation
}
\examples{
\dontshow{if (.Platform$OS.type != "windows") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
qry <- system.file("extdata", "toy.cg", package = "MethScope")
msk <- system.file("extdata", "toy.cm", package = "MethScope")
res <- GenerateInput(qry, msk)
\dontshow{\}) # examplesIf}
}
