% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelTraining.R
\name{Input_training}
\alias{Input_training}
\title{Train XGBoost model to predict cell type}
\usage{
Input_training(
  summary_results,
  cell_type_label,
  number_patterns = 1000,
  cross_validation = FALSE,
  xgb_parameters = list()
)
}
\arguments{
\item{summary_results}{a wide cell by pattern matrix generated from GenerateInput function}

\item{cell_type_label}{a vector of the corresponding cell type label for each row of the summary results}

\item{number_patterns}{a numeric value to indicate number of patterns to be used (Default: 1000)}

\item{cross_validation}{a boolean varaible whether to perform cross_validation to obtain the best hyper parameters for the model}

\item{xgb_parameters}{an optional list for xgb model parameters provided by the user}
}
\value{
the xgb model trained
}
\description{
Train XGBoost model to predict cell type
}
