% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PretrainedModels.R
\name{Liu2021_MouseBrain_P1000}
\alias{Liu2021_MouseBrain_P1000}
\title{Load the Liu et al. mouse brain pretrained model}
\usage{
Liu2021_MouseBrain_P1000()
}
\value{
A MethScopeModel object for use with \code{\link{PredictCellType}}.
}
\description{
Load the Liu et al. mouse brain pretrained model
}
