% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictCellType.R
\name{nnls_deconv}
\alias{nnls_deconv}
\title{Estimate cell type relative proportion}
\usage{
nnls_deconv(ref, mixture_matrix, number_patterns = 1000)
}
\arguments{
\item{ref}{An imputed wide cell by pattern matrix generated from GenerateInput function using reference Pseudobulk}

\item{mixture_matrix}{An imputed wide cell by pattern matrix generated from GenerateInput function}

\item{number_patterns}{a numeric value to indicate number of patterns to be used (Default: 1000)}
}
\value{
A cell type by cell matrix showing the relative cell type proportion estimate for each cells
}
\description{
Estimate cell type relative proportion
}
