% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeOutput.R
\name{PlotF1}
\alias{PlotF1}
\title{Generate F1 score barplot for each class}
\usage{
PlotF1(prediction_result, actual_label)
}
\arguments{
\item{prediction_result}{Prediction result from PredictCellType}

\item{actual_label}{Ground truth cell label}
}
\value{
A ggplot2 object.
}
\description{
Generate F1 score barplot for each class
}
