% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeOutput.R
\name{PlotUMAP_fixedwindow}
\alias{PlotUMAP_fixedwindow}
\title{Generate UMAP for the final prediction based on fixed window eg.100kb bin widows}
\usage{
PlotUMAP_fixedwindow(
  query_fn,
  knowledge_fn,
  prediction_result,
  n_component = 30,
  seed = 123,
  ...
)
}
\arguments{
\item{query_fn}{File path to query .cg}

\item{knowledge_fn}{File path to 100bk bins window or reference pattern}

\item{prediction_result}{Prediction result from PredictCellType}

\item{n_component}{Number of PCA components to use (Default: 30)}

\item{seed}{A number for random seed (Default: 123)}

\item{...}{Additional arguments passed to \code{uwot::umap} (e.g., \code{n_neighbors}, \code{metric}).}
}
\value{
A list of two ggplot2 UMAP object.
}
\description{
Generate UMAP for the final prediction based on fixed window eg.100kb bin widows
}
