% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictCellType.R
\name{PredictCellType}
\alias{PredictCellType}
\title{Predict cell type annotation from the trained model}
\usage{
PredictCellType(bst_model, predictMatrix, smooth = FALSE, KNeighbor = 5)
}
\arguments{
\item{bst_model}{The boosting model trained from ModelTrain}

\item{predictMatrix}{A wide cell by pattern matrix generated from GenerateInput function}

\item{smooth}{A Boolean variable to indicate whether smooth the matrix (Default: FALSE)}

\item{KNeighbor}{number of knn neighbors to use for smoothing (Default: 5)}
}
\value{
A cell by cell type matrix with confidence score and labeled cell type.
}
\description{
Predict cell type annotation from the trained model
}
\examples{
\dontrun{
# Run from the root of a cloned zhou-lab/MethScope repository.
input_pattern <- GenerateInput("inst/extdata/example.cg",
                               "inst/extdata/mm10_Liu2021.cm")
model <- Liu2021_MouseBrain_P1000()
prediction <- PredictCellType(model, input_pattern)
}
}
