% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictCellType.R
\name{smooth_matrix}
\alias{smooth_matrix}
\title{Smooth cell by pattern matrix to reduce noise}
\usage{
smooth_matrix(predictMatrix, KNeighbor = 5)
}
\arguments{
\item{predictMatrix}{A wide cell by pattern matrix generated from GenerateInput function}

\item{KNeighbor}{Number of knn neighbors to use for smoothing (Default: 5)}
}
\value{
A wide cell by pattern matrix after smoothing and knn graph
}
\description{
Smooth cell by pattern matrix to reduce noise
}
