% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PretrainedModels.R
\name{Zhou2025_HumanAtlas_P1000}
\alias{Zhou2025_HumanAtlas_P1000}
\title{Load the Zhou et al. human atlas pretrained model}
\usage{
Zhou2025_HumanAtlas_P1000()
}
\value{
A MethScopeModel object for use with \code{\link{PredictCellType}}.
}
\description{
Load the Zhou et al. human atlas pretrained model
}
