% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateInput.R
\name{GenerateInput}
\alias{GenerateInput}
\title{Generate pattern level data for cell type annotation}
\usage{
GenerateInput(query_fn, knowledge_fn)
}
\arguments{
\item{query_fn}{File path to query .cg}

\item{knowledge_fn}{File path to pattern file .cm}
}
\value{
A cell by pattern matrix.
}
\description{
Generate pattern level data for cell type annotation
}
\examples{
# Use the example.cg file included in the package
reference_pattern <- system.file("extdata", "Liu2021_MouseBrain.cm", package = "MethScope")
example_file <- system.file("extdata", "example.cg", package = "MethScope")
result <- GenerateInput(example_file,reference_pattern)
}
