% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateInput.R
\name{GenerateInput}
\alias{GenerateInput}
\title{Generate pattern level data for cell type annotation}
\usage{
GenerateInput(query_fn, knowledge_fn)
}
\arguments{
\item{query_fn}{File path to query}

\item{knowledge_fn}{File path to pattern file}
}
\value{
A cell by pattern matrix.
}
\description{
Generate pattern level data for cell type annotation
}
