% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateInput.R
\name{GenerateReference}
\alias{GenerateReference}
\title{Generate reference pattern file}
\usage{
GenerateReference(binary_file, min_CG = 50, output_path = "./patterns.txt")
}
\arguments{
\item{binary_file}{File path from the output of the bash script}

\item{min_CG}{The minimum CpG # a pattern must have (Default: 50)}

\item{output_path}{File path to store the output txt file (Default: current path)}
}
\value{
NULL, write out a patterns.txt file
}
\description{
Generate reference pattern file
}
