% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeOutput.R
\name{imputeRowMean}
\alias{imputeRowMean}
\title{Impute missing value for 100K window matrix}
\usage{
imputeRowMean(mtx)
}
\arguments{
\item{A}{cell by 100K window data frame with missing values}
}
\value{
A cell by 100K window data frame with imputed values
}
\description{
Impute missing value for 100K window matrix
}
