% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeOutput.R
\name{PlotConfusion}
\alias{PlotConfusion}
\title{Generate confusion table for the final prediction}
\usage{
PlotConfusion(prediction_result, actual_label)
}
\arguments{
\item{prediction_result}{Prediction result from PredictCellType}

\item{actual_label}{Ground truth cell label}
}
\value{
A ggplot2 confusion table object.
}
\description{
Generate confusion table for the final prediction
}
