% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeOutput.R
\name{PlotUMAP}
\alias{PlotUMAP}
\title{Generate UMAP for the final prediction based on cell patterns}
\usage{
PlotUMAP(predictMatrix, prediction_result, n_component = 30, seed = 123, ...)
}
\arguments{
\item{predictMatrix}{a wide cell by pattern matrix generated from GenerateInput function}

\item{prediction_result}{Prediction result from PredictCellType}

\item{n_component}{Number of PCA components to use (Default: 30)}

\item{seed}{A number for random seed (Default: 123)}

\item{...}{Additional arguments passed to \code{uwot::umap} (e.g., \code{n_neighbors}, \code{metric}).}
}
\value{
A list of two ggplot2 UMAP object.
}
\description{
Generate UMAP for the final prediction based on cell patterns
}
