% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictCellType.R
\name{PredictCellType}
\alias{PredictCellType}
\title{Predict cell type annotation from the trained model}
\usage{
PredictCellType(bst_model, predictMatrix)
}
\arguments{
\item{bst_model}{boosting model trained from ModelTrain}

\item{predictMatrix}{a wide cell by pattern matrix generated from GenerateInput function}
}
\value{
A cell by pattern matrix.
}
\description{
Predict cell type annotation from the trained model
}
