% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictCellType.R
\name{PredictCellType}
\alias{PredictCellType}
\title{Predict cell type annotation from the trained model}
\usage{
PredictCellType(bst_model, predictMatrix, smooth = FALSE, KNeighbor = 5)
}
\arguments{
\item{bst_model}{The boosting model trained from ModelTrain}

\item{predictMatrix}{A wide cell by pattern matrix generated from GenerateInput function}

\item{smooth}{A Boolean variable to indicate whether smooth the matrix (Default: FALSE)}

\item{KNeighbor}{number of knn neighbors to use for smoothing (Default: 5)}
}
\value{
A cell by cell type matrix with confidence score and labeled cell type.
}
\description{
Predict cell type annotation from the trained model
}
\examples{
# Use the example.cg file included in the package
reference_pattern <- system.file("extdata", "Liu2021_MouseBrain.cm", package = "MethScope")
example_file <- system.file("extdata", "example.cg", package = "MethScope")
result <- GenerateInput(example_file,reference_pattern)
prediction_result <- PredictCellType(MethScope:::Liu2021_MouseBrain_P1000,result)
}
